Unclutter cli.py a bit by moving subtitle stuff to util.py
